fix coverity "bug" 184   (false positive)
the contract for wiretap ..._read() functions are that the *err pointer MUST always
be provided as a valid pointer and not a NULL pointer.


ber.c contained a spurios and redundant check for the pointer being non-NULL wich triggered coverity to detect this as a false positive.

this redundant test if the pointer is NULL has been revoved.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@18492 f5534014-38df-0310-8fa8-9805f1628bb7