coverity   187
i dont think this can trigger in the current code   but better safe than sorry  and it safeguards against future changes that might break this requirement.


dereferencing a nullpointer would cause crash of wireshark   in case a new bug is introduced making it possible for this variuable to be null.





git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@18543 f5534014-38df-0310-8fa8-9805f1628bb7