rename some inline functions from se_ to emem_ prefix as a first step to generalize teh tree management and to use trees with different storage scope without too much code duplication.
it would be useful with a tree that had indefinite storage instead of the emem functions which commonly have ep or se storage scope.

indefinite storage scope would be useful for example for managing a global and static set of well known guid to name mappings(not yet implemented)    and also for
oid to name mappings.

 


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@18886 f5534014-38df-0310-8fa8-9805f1628bb7