make creation of subtrees used for the _array tree type use the same allocation scope as its parent  and thus become allocation scope agnostic
change all accessor functions to be defines to the emem_tree_ functions.



now   to create a tree with a different scope we only need to create a new
..._tree_create() function   and set up the appropriate defines


(it was a mistake to call the functions   se_tree_create   and se_tree_create_non_persistent,     they should be the other way around    i.e.  se_tree_create_persistent   and se_tree_create )




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@18895 f5534014-38df-0310-8fa8-9805f1628bb7