from Angel de Juan
make mms bitstrings easier to read


--
Hi, this is my new mms with some changes in order to display some data in
binary instead of hex.

Excuse me because is not a diff from the repository, but I don't know how to
create a diff.

I also have the compiled packet-mms.c and packet-mms.h if you want it only
mail me.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@18974 f5534014-38df-0310-8fa8-9805f1628bb7