from bryant eastham
Also, there is still an outstanding issue regarding the default use of
the "media" dissector. The way it is currently coded there is no way to
have a heuristic decoder when a content-type header is specified.


In this way if there is a decoder for a specific content-type then it
will be used, then the heuristic decoders have a chance, and finally the
default of either the media-type decoder of the http_payload decoder.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@19208 f5534014-38df-0310-8fa8-9805f1628bb7