From G.F. aka Gena01 I have put together a patch for YMSG packet dissector. This is based on my own code and service lists (this should match Gaim and Kopete service lists). This new code should bring the code up to par to most of the known services. Which should cover up to Yahoo 7.x or most of it.
I have also setup a new set of constants which are specific to YMSG packets. These are the types that I've
seen in miranda network logs and they should reveal more information. The other constants are mostly for buddy
statuses and need nor apply to the YMSG header.  I have left them in the code (for now). These constants are
currently used in my own code.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@19255 f5534014-38df-0310-8fa8-9805f1628bb7