"Pluggable custom makefile" A custom plugin (not yet to be released to the public) can be added to Makefile.nmake easily. But this has a drawback: The modified Makefile.nmake appears as a modified file in the commit list and in generated patches (this annoyed me for years).
To solve this I've added a plugin like mechanism which calls a Custom.nmake file for each target (if the file is existing). I've also added a Custom.nmake.example for a quick start.

I don't know how to do this for the Makefile.am, so I leave this as an exercise for the interested reader ;-)

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@19377 f5534014-38df-0310-8fa8-9805f1628bb7