Add a WIRESHARK_RUN_FROM_BUILD_DIRECTORY environment variable on UN*X; if set, and if the program isn't running with additional privileges, it'll treat the directory in which the program is found as the data directory.
If, on Windows, the version-number subdirectory of {data
directory}\plugins doesn't exist (which is assumed to mean that the
program is being run from the build directory), or if, on UN*X,
WIRESHARK_RUN_FROM_BUILD_DIRECTORY is set, the plugin directory is the
"plugins" subdirectory of the data directory, and all subdirectories of
that directory are scanned for plugins, as the "plugins" subdirectory of
the build directory contains subdirectories for the plugins; this means
that if we're running from the build directory, we'll find the plugins
we built in the build tree.

When generating the wireshark-filter man page, run tshark with
WIRESHARK_RUN_FROM_BUILD_DIRECTORY set, so it uses the plugins from the
build to generate the list of filters.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@20261 f5534014-38df-0310-8fa8-9805f1628bb7