From Sebastien Tandel:
    Here is an updated patch for proto_tree_add_item and the
range_string structure. The new macro RVALS() can be used as the macro
VALS() in the declaration of your hf_register_info with another
structure (range_string).  Be aware that you *have to* ORed the value of
the field display with BASE_RANGE_STRING constant and it can 'only' be
used with FT_(U)INT* types in a header_field_info.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@20805 f5534014-38df-0310-8fa8-9805f1628bb7