From Peter Johansson:
I have provided a patch for a crash that occurs in uat_gui.c if from the
"Edit preferences..." button when you choose Protocols -> DLT_USER ->
the "Edit..." button -> the "New" button -> the "Cancel" button.
The crash happens because dd gets freed prior to traversing its internal
structure, hence I moved the call to g_free for dd beneath the traversal
of the dd internals.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@20813 f5534014-38df-0310-8fa8-9805f1628bb7