From Florent Drouin:
This patch provide a new function to decode messages when several ASN1 encoding can be used.
This is the case, for example, when a same message has different encoding according to the MAP version, or in case of ASN1 encoder optimization.
 At the same time, I did remove the configuration variable "old_gsm_map_version", which is taken into account in the patch.
And likewise, for all the messages defined with the ASN1 sequence "[3] SEQUENCE", this function is called for the decoding.

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@20878 f5534014-38df-0310-8fa8-9805f1628bb7