Fix bug #1409:
There is an error in the page:
http://wireshark.org/docs/wsug_html_chunked/ChCapCaptureFilterSection.html

According to Example 4.2, the filter 'tcp port 23 and not host 10.0.0.5'
captures all telnet traffic not from 10.0.0.5. But this filter also discards
traffic to 10.0.0.5. Either you use this filter to capture all telnet traffic
not to and not from 10.0.0.5 or you use filter 'tcp port 23 and not src host
10.0.0.5'.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@20955 f5534014-38df-0310-8fa8-9805f1628bb7