All private data for a capture in a Wiretap module must be per-file, not static to the module.
Add the older(?) ID tag for MPEG audio.

Just use the ID at the beginning to identify MPEG audio files; don't
check the file any further.

If the read of the magic number doesn't work, get the error, and, if
there is no error (i.e., it's a short read), just return 0 (meaning "no
error, but this isn't that type of file).

Similarly, if the magic number doesn't match, just return 0, so other
types of file are tried.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@21192 f5534014-38df-0310-8fa8-9805f1628bb7