From Mike Duigou:
The enclosed patch corrects a problem where jxta elements were being added to the protocol tree for segments that did not contain complete jxta frames. This patch ensures that the jxta proto elements are only added those the segments that end a complete, assembled jxta frame.

The patch has been fuzz tested with a broad selection of jxta captures and ran successfully overnight for over 4000 iterations.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@21305 f5534014-38df-0310-8fa8-9805f1628bb7