From Sake Blok:
wireshark SHOULD be able to filter on multiple hf's with the
same field-name, BUT there is a little bug in the code. I have pinpointed it to
the following in epan/dfilter/dfvm.c:
...
It actually loops through all the hf's with the same name, but only checks
against the original (first) hf.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@21372 f5534014-38df-0310-8fa8-9805f1628bb7