With at least some compilers, "enum QUALITY_TYPE" is unsigned, and the compiler warns about checking whether it's >= 0, as it always is.
Don't do that check - we don't use "type" until the switch statement,
and GCC should issue a warning if not all valid "enum QUALITY_TYPE"
values are handled in the switch statement.

Sort the items in the switch statement in the order in which they appear
in the enum definition, so it's clearer that it covers all values.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@21890 f5534014-38df-0310-8fa8-9805f1628bb7