From Martin Warnes:
Attached is a small patch that adds a preference option to allow the SQL
Query string to be appended to the INFO column display. This makes life
a little bit easier when scanning a trace to find the packet associated
with a specific query. I thought it was better to implement it as a
preference option with default set FALSE as it can clutter up the
display for applications that generate large amounts of queries.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@22130 f5534014-38df-0310-8fa8-9805f1628bb7