disabling sid snooping in the code
this module was never really finished and was therefore semi-useless.
disabled now since the change to samr broke it  and the real fix to 
unbreak it would be a substantial rewrite of it
(and adding a nice gui so one can see which sids are mapped...)

disabled it   but did not remove it since i do think the module is 
useful   IF it is completed with a lot more operations.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@22271 f5534014-38df-0310-8fa8-9805f1628bb7