From Kevin A. Noll via private bug #1683:
This is a major re-write of the WLCCP version 0xc1 dissector. It now is
able to dissect many more WLCCP message types and the TLVs that may be
included in a message. Placeholders are left for message types and
TLV types that we do not currently have enough data to engineer a
dissector.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@22330 f5534014-38df-0310-8fa8-9805f1628bb7