From Mike Duigou <wireshark@duigou.org> (bug 1698):
Enclosed is a patch with fixes problems with the desegmentation of chunked HTTP
message bodies.

...

Changing the value of pinfo->desegment_len to DESEGMENT_ONE_MORE_SEGMENT rather
than the current 1 or 2 seems to make everything work. (I'm kind of confused as
to why the headers resassembly code already used DESEGMENT_ONE_MORE_SEGMENT and
the body reassembly code did not).



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@22570 f5534014-38df-0310-8fa8-9805f1628bb7