Fix problems found by Beyond Security:
Instead of overflowing an unsigned int when determining the number of
items in a range, use a signed int and check for a negative value.  Make
sure our offset increments as we step through each item.  This should 
avoid large/inifinite loops.

Fix the size of hf_dnp3_al_range_stop32.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@22811 f5534014-38df-0310-8fa8-9805f1628bb7