The dissector packet-dcp.c now fails to compile on MacOS X (gcc 4.0.1/Apple).  Change G_INLINE_FUNC back to inline per GLib documentation (also remove recommendation in developer's guide to not use inline):
http://library.gnome.org/devel/glib/unstable/glib-Miscellaneous-Macros.html#id2571572

G_INLINE_FUNC

#define             G_INLINE_FUNC
This macro is used to export function prototypes so they can be linked with an external version when no inlining is performed. The file which implements the functions should define G_IMPLEMENTS_INLINES before including the headers which contain G_INLINE_FUNC declarations. Since inlining is very compiler-dependent using these macros correctly is very difficult. Their use is strongly discouraged.

This macro is often mistaken for a replacement for the inline keyword; inline is already declared in a portable manner in the glib headers and can be used normally.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@22980 f5534014-38df-0310-8fa8-9805f1628bb7