This patch is a remake of SVN 22762:
When doing TCP_SEQ analysis, if the packet is a SYN, then it's 
not a lost packet but the tcp ports are being reused. This is often 
seen in load-balanced environments where client ports are preserved 
on the server-side.

This time it is fixed by creating a new conversation whenever a
new SYN is received for an existing conversation. This fixes the
following:

- bug 1680: Error in TCP Sequence number analysis
- TCP-conversation timestamps for new TCP-sessions with the addresses 
  and ports as a previous TCP-conversation in the trace-file.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@23299 f5534014-38df-0310-8fa8-9805f1628bb7