Add two sentences clarifying the license statement on pidl:
As a notable exception the pidl utility at tools/pidl is covered by a
"GPL version 3 or later" LICENSE. Note that only the tool itself
is covered by this license, not the source code generated by it. The
pidl authors do not consider that code a derived work of pidl.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@23387 f5534014-38df-0310-8fa8-9805f1628bb7