This patch makes it possible to disable individual coloring rules without having to delete them. The patch has been tested on Fedora-7 with GTK+ 1.2.10 and GTK+ 2.10.11.
Since I don't know how to use "strikethrough" in clists in GTK1
there is a little difference in how the disabled coloring rules
are displayed. In GTK2 they are striked through and in GTK1
they are shown in lightgrey on a white background.

Any info on how to use strikthrough in clists within GTK1 is more
than welcome :-)



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@23421 f5534014-38df-0310-8fa8-9805f1628bb7