Clean up indentation some more.
Don't pass an argument if the argument is overwritten before being used;
make it a local variable instead.

Combine the header and contents of an object into a single tree, as per
a suggestion in bug 1870.

Clean up some flag displays, and put fields into the protocol tree in
the order in which they appear in the packet.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@23960 f5534014-38df-0310-8fa8-9805f1628bb7