From Thomas Morin: I have written the attached patch to improve the userfriendly-ness of the summary displayed in the last column of the packet list for IGMP packets.
The current code only tells the user the type of packet (REPORT, LEAVE, QUERY),
but doesn't include any information about what multicast group is joined to or
left.   Worse, for IGMPv3 Reports the user cannot tell quickly if the multicast
group is joined-to or left, without having to look at the inside of the packet
to know the type of report (include all, include none, exlude none, exclude
all, etc.).

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@24190 f5534014-38df-0310-8fa8-9805f1628bb7