Check in changes originally checked in as SVN revision 24308 (new column type: Custom) that were backed out in SVN revision 24309.
Changes since that revision include a reworking of the handling of the
cfile/cinfo variables in epan/column-utils.c, addition of three new
functions to libwireshark.def and a bug fix to prevent a crash when no
custom columns were not in use.

Compilation verified locally on MacOS X, Linux and Windows.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@24317 f5534014-38df-0310-8fa8-9805f1628bb7