Put back the use of g_ascii_strcasecmp(), so we don't get rudely surprised by, for example, "i" not comparing equal to "I" in Turkish locales.
Make gMessageType local - it's not used as a global variable.  Also,
make it a pointer to const, so that we don't have to cast away the
constness.

dissect_giop_fragment() doesn't use the "header" argument, so get rid of
it.

Clean up indentation a bit.

The first element of a value_string is the value, and the second
argument is the string; fix up the initialization of the table of code
set values.

Fix a proto_tree_add_text() call to match the format and the arguments.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@24411 f5534014-38df-0310-8fa8-9805f1628bb7