Mark a local function static rather than extern. Get rid of an unused variable.
Was going to use proto_register_prefix(), but am not sure if its appropriate.
Even with tree == NULL, dictionary items are used to generate info column.
Loading the dictionary lazily would mean doing a lot of work as part of dissecting
the first diameter frame, which could decrease our chances of capturing frames that
arrive very soon after that one...

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@24784 f5534014-38df-0310-8fa8-9805f1628bb7