Put a space between the number of bytes (for a file < 1024 bytes) and "Bytes".
"Interface description" vs. "interface name" isn't Windows vs. UN*X,
it's "has a description" vs. "doesn't have a description" - the lack of
descriptions in UN*X is a result of

	1) lack of a mechanism to supply descriptions in some UN*Xes

and

	2) lack of code in libpcap to get those descriptions in some
	   other UN*Xes

and I need to work on fixing 2) (I think it can be dredged out of
FreeBSD 6 and 7, I seem to remember current OpenBSD supplying them, and
for OS X we might as well cheat the way the network preference pane does
and just call en0 the main Ethernet and en1 the AirPort interface).

Use an adverb rather than an adjective to modify a noun, as per Chris
Maynard's suggestion.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@25089 f5534014-38df-0310-8fa8-9805f1628bb7