Make the preference settings for the dissector to use for various ERF link-layer types static.
If the preference is set to "raw data" for any of those types, display
the raw data with an indication that the preference in question has been
set to "raw data", so people don't just wonder why ERF files aren't
working right any more.  (See bug 2641; I had the same surprise when I
tried it on some ERF captures I have.)

Pull the dissect_erf_header() code into dissect_erf() - it's dissecting
the *payload*, not the *header*.

Fill in the Info column with the record type.

When using tvb_new_subset() to chop a header off of a tvbuff, just
specify lengths of -1, so we go all the way to the end.

Clean up the Infiniband dissector call.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@25608 f5534014-38df-0310-8fa8-9805f1628bb7