Replace the old "erfatm" preference for the ERF dissector with an "aal5_type" dissector, which offers only "guess the traffic type" and "LLC multiplexed" as options, defaulting to "guess the type".
Add a separate preference to control whether to treat single ATM cells
as raw data or as the first cell of an AAL5 PDU (and dissecting them as
short AAL5 PDUs).

Don't reach inside the tvbuff to get the data and the length; use
tvb_length() and tvb_get_ptr().

Pass the data *after* the AAL5 header to the "guess the traffic type"
routine.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@25736 f5534014-38df-0310-8fa8-9805f1628bb7