Link libwireshark and wiretap against libwsutil.  That was the original intent of adding libwsutil but somehow I missed it/got it wrong.  This should solve https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=1677#c18
Also remove the _DEPENDENCIES lines in epan and wiretap (as was done in the
top-level Makefile in rev 25930) so that automake will automatically figure out
the dependencies for us.

Since the those 2 libraries now link against libwsutil, don't make every
executable link against the library.  (If this works I think we can
significantly trim the list of libraries the executables link against and just
let the libraries pull in what they need--which is, apparently, the point of
the --as-needed flag: http://www.gentoo.org/proj/en/qa/asneeded.xml ).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@26218 f5534014-38df-0310-8fa8-9805f1628bb7