From Colin O'Flynn: The idea is that there is now some hardware that can put 802.15.4 frames over Ethernet. To do so, the 802.15.4 frames are wrapped in an Ethernet frame, with the Ethertype set to a value indicating the payload is 802.15.4.
Since there is no official ETHTYPE designated by the IEEE, the number 0x809A
is used in this code. However a preference is added to the "IEEE 802.15.4" type
in the preference dialog allowing you to change this ethtype to something else.

The hardware for those interested is the Atmel Raven USB Stick.

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@26352 f5534014-38df-0310-8fa8-9805f1628bb7