From Joe Eykholt: The dissector for FIP (FCoE (Fibre Channel over Ethernet) Initialization Protocol) needs to be updated for recent changes and proposed changes by T11.
The original version was done when the protocol was fairly close to being
ratified, but it is very very close now.  The letter ballot has gone out.
The only part of the attached patch that isn't on the current ballot is the
VLAN descriptor.  I will update that later if it turns out to be changed before adoption.

The changes are as follows:
The SWWN descriptor is now called a Name descriptor.
The FWWN descriptor is now called a Fabric descriptor and has two new fields,
VFID (virtual fabric ID) and FC-MAP OUI value.
The MRS descriptor is now called FCoE Size.
The FKA descriptor units changed from seconds to milliseconds.
The VLAN request and reply ops have been added.
A VLAN descriptor type has been added.

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@26629 f5534014-38df-0310-8fa8-9805f1628bb7