Add (preliminary, and a bit hacky) support for interface-type icons on OS X.
Add some comments to the interface-type icon code, and, when looking for
certain text in the description on Windows, look in the interface
description supplied by WinPcap rather than the user-supplied
description (if we want to look in the latter, we presumably want to
look in both, so the user giving a wireless interface a name they like
won't break the code that checks for wireless interfaces).  There's
probably *some* NDIS OID that can give you the information you want -
hopefully, not an NDIS 6-only OID).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@27345 f5534014-38df-0310-8fa8-9805f1628bb7