The reason why the expert_item value is null when performing a retap is that a retap doesn't build the protocol tree unless there's a tap with a filter; presumably "the expert_item value" refers to the item to which expert information is being attached.
However, if there's no protocol item, the only thing
expert_add_info_format() doesn't do is set the severity and group for a
protocol item, but the expert info composite dialog doesn't appear to
look at that.

Thus, it appears we can just use cf_retap_packets(); do so, as it's
faster.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@28270 f5534014-38df-0310-8fa8-9805f1628bb7