From Ben Winslow via bug 3535:
Because Lucent/Ascend equipment will sometimes omit the hex dump for a packet
or send two headers followed by two hex dumps, Wireshark needs to be very
lenient when parsing a Lucent/Ascend trace.  On a busy access server, a packet
like this is pretty likely to appear within a few minutes.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@28749 f5534014-38df-0310-8fa8-9805f1628bb7