Don't use SIGUSR1 to tell dumpcap to exit, use SIGINT: SIGINT is traditionally used for this purpose and using it also prevents the 2 signals the child gets: 	- the user's Ctrl-C (which is sent as a SIGINT to both *shark and its 	  child dumpcap) 	- the signal *shark generates to shut down the child
from colliding (and running 2 signal handlers in the child).

It might be possible for tshark to not send the signal at all when it gets
SIGINT, but it doesn't do any harm now.

Also, do not call g_log() within the signal handler: doing so can cause
aborts (if g_log is being called by the process when the signal comes, the
2nd entrance into g_log is detected as a recursion).

This fixes https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=2767


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@29881 f5534014-38df-0310-8fa8-9805f1628bb7