From Don Chirieleison via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3914 :
SDNVs are theoretically unlimited in size.  The value of most SDNVs in the
Bundle Protocol is practically limited to far less than a 32 bit number. The
initial dissector included only 1 SDNV evaluation routine which returned a 32
bit number. SDNV fields that evaluated to greater than a 32 bit number were
considered in error. One BP implementation chose to add some syntax to one of
the SDNV fields that extends it to more than 32 bits. The patch included here
adds an evaluation routine that will return a 64 bit number. That routine is
called to evaluate the field where it makes sense to have a value in excess of
32 bits.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@29954 f5534014-38df-0310-8fa8-9805f1628bb7