From Peter Harris via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=4481 :
The XKEYBOARD extension is a pain in the... so XCB had to grow new ways to
describe the protocol.

This patch adds support to the Wireshark dissector, so it can successfully
process the XCB description of the XKEYBOARD extension.

Without this patch, "make x11-dissector" chokes on the current xcbproto from
git.

Also update the script to deal with the new mesa directory structure.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@32520 f5534014-38df-0310-8fa8-9805f1628bb7