From Michael Bernhard via bug 3398 with minor fixups:
I've created a ASN.1 dissector for the IEC 61850 Sampled Values protocol. It
dissects ethernet frames of the IEC 61850-9-2LE specification form the UCA
International User Group.

There is also a new TAP for tshark (-R sv) which extracts the important
information of the frame and allows to create plots (with external tools) of
the sampled values.

I've developed under Linux (Ubuntu 8.10) but everything should be in place for
successful compilation under Windows.

It would be great if this dissector could be included in wireshark. I'm looking
forward for your comments.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@33039 f5534014-38df-0310-8fa8-9805f1628bb7