Sorry, I forgot that 8.6. UDVM Cycles says
	Note that the number of UDVM cycles MUST NOT be increased if a
	request for additional compressed data fails.

so we *shouldn't* increase the cycle count further when
SIGCOMP_INSTR_INPUT_BYTES fails.  That does *NOT* of course, mean that
we shouldn't increase the cycle count by 1 for a failed INPUT-BYTES
instruction - that would leave UDVM vulnerable to infinite loops (as per
bug 4826), and I *really* doubt that was their intent; presumably, it
means it should not be increased *by the number of cycles for the
additional data*:

	Additionally, if the UDVM successfully requests n bits of
	compressed data using one of the INPUT instructions then the
	number of available UDVM cycles is increased by n *
	cycles_per_bit once the instruction has been executed.

if the attempt to get that additional data fails.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@33065 f5534014-38df-0310-8fa8-9805f1628bb7