From Cal Turney via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=4831 :
Add the data read/write length to the NFS tree so it is filterable.

From me: don't bother incrementing the offset just to decrement it again.
Change the hf info a bit.

(Ideally the RPC dissector would add the length to the tree not as a text
item; that is left for future work.)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@33101 f5534014-38df-0310-8fa8-9805f1628bb7