If we get an error reading the capture file, don't give up - run the taps, etc., so we at least print the statistics for the packets we *could* read.
Also, if we're printing packet data, and the standard output and
standard error are going to the same place, before printing the error
message for that error, flush the standard output and print a newline to
the standard error, so the error appears after all the packet data, and
has a newline separating it from the packet data.

This fixes bug 4845.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@33124 f5534014-38df-0310-8fa8-9805f1628bb7