From Turney Cal as a followup to rev 33841 and bug 5038:
I tested your update ("wireshark -r foo.pcap" from a cmd window) and 
found that it did not use the current folder for the next Open dialog 
and didn't save the folder in recent.  The attached patch prepends the
filename with the current folder.

From me: don't leak memory when doing it.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@33863 f5534014-38df-0310-8fa8-9805f1628bb7