Disable the "File Set" menu while capturing, because changing file while capturing causes an assert in cf_reset_state().
We still get this assert if the "List Files" dialog is opened before we start
capture and we change file during capturing.  Maybe we have to close this
dialog when starting capture?  Or maybe disable all files except the one
being captured to?  Then we also may have to update the "Last Modified" and 
"Size" columns.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@34234 f5534014-38df-0310-8fa8-9805f1628bb7