From Roberto Morro: Fix for Bug #5628
This patch (against revision 35716) resolves the following bug:
[introduced in patch attached to Bug #5518]
"Dissector bug, protocol RSVP, in packet 78:
More than 1000000 items in the tree -- possible infinite loop"

Actually the bug is resolved by the "return" statement that was erased by
mistake. The patch fixes another issue: the code didn't take care of the fact
that TLVs must be padded to a length multiple of 4.

See:
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5518
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5628

From me:
 Change two bytestring_to_str(tvb_get_ptr(...),...) to tvb_bytes_to_str(tvb,...)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@35722 f5534014-38df-0310-8fa8-9805f1628bb7