For FT_{ABSOLUTE,RELATIVE}_TIME, make the rest of the encoding orthogonal to the byte order.
This means that we can't just test for a non-zero encoding to determine
whether the format is big-endian or little-endian when we set the
field's endianness flag; instead, for the types where we accept any
non-zero value as meaning "litle-endian", map it to ENC_LITTLE_ENDIAN.

When we use ENC_TIME_NTP, OR in the byte order flag.  While we're at it,
in the dissectors that used ENC_TIME_NTP, update all the other encoding
items in proto_tree_add_item() calls to use the appropriate ENC_ value.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@35841 f5534014-38df-0310-8fa8-9805f1628bb7