Update and expand some comments.
In the end-of-stream code, when we're checking the CRC and length, don't
check the CRC or length if we failed to read them, and don't check the
length if the CRC is bad.

We define O_BINARY as 0 on UN*X in <wsutil/file_util.h>, so we don't
need to avoid using it on UN*X.

In file_gets(), check for delayed errors.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@36590 f5534014-38df-0310-8fa8-9805f1628bb7