Make the packet count an unsigned value, as frame numbers are unsigned.
Make the loops that scan through all the packets do so by frame number,
to abstract away the "next" and "previous" pointers in the frame_data
structure.  Add a routine to cfile.c to map frame numbers to frame_data
structures, and put in some special case handling so scanning forward or
backward through the packets is O(N) rather than O(N^2).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@36846 f5534014-38df-0310-8fa8-9805f1628bb7